Cleaned NavBar component
