+++ src/pages/components/NavigationComponent/RnNavBar.tsx
@@ -28,97 +28,15 @@ import {
   Route,
   Link,
 } from "react-router-native";
+import { routes } from "../../../routeConfig";
+import RenderFirstLevelRoutingComponent from "../RenderComponent/RenderFirstLevelRoutingComponent";
+import RenderSecondLevelRoutingComponent from "../RenderComponent/RenderSecondLevelRoutingComponent";
 
 import ExpandableComponent from "./RnExpandableComponnent";
 
 const Home = () => (
   <Text style={styles.header}>Default First Level Route 1 to be show</Text>
 );
-
-function RenderFirstLevelRoutingComponent({ routes }) {
-  const history = useHistory();
-
-  return (
-    <View>
-      <View style={styles.nav}>
-        <Row>
-          <View
-            style={{
-              flex: 1,
-              borderWidth: 2,
-            }}
-          >
-            <Text
-              style={{
-                margin: 10,
-              }}
-            >
-              First Level of Route 2 :-
-            </Text>
-            <View
-              style={{
-                marginLeft: 200,
-                marginRight: 200,
-                marginBottom: 50,
-              }}
-            >
-              <Button
-                title="Click here to View the Second Level Route Below"
-                onPress={() => {
-                  console.log("First Row Button Clicked");
-                  history.push("/demo_route2/list");
-                }}
-              />
-            </View>
-          </View>
-        </Row>
-      </View>
-
-      {routes.map((route, i) => (
-        <RouteWithSubRoutes key={i} {...route} />
-      ))}
-    </View>
-  );
-}
-
-function RenderSecondLevelRoutingComponent(props) {
-  return (
-    <View>
-      <Row>
-        <View
-          style={{
-            margin: 10,
-            borderWidth: 2,
-          }}
-        >
-          <Text>Got 2nd Level Route 2</Text>
-        </View>
-      </Row>
-    </View>
-  );
-}
-
-// JSON ROUTES DEFINED
-const routes = [
-  // {
-  //     path: "/",
-  //     component: Home
-  // },
-  {
-    path: "/demo_route2",
-    component: RenderFirstLevelRoutingComponent,
-    routes: [
-      {
-        path: "/demo_route2/list",
-        component: RenderSecondLevelRoutingComponent,
-      },
-      // {
-      //     path: "/tacos/cart",
-      //     component: Cart
-      // }
-    ],
-  },
-];
 
 function RouteWithSubRoutes(route) {
   return (
